progress

+++ src/index.js
@@ -1,5 +1,7 @@
+import colors from "ansi-colors";
+import { timeout } from "thyming";
 import { getClient } from "./discord";
-import { updateUser } from "./users";
+import { updateUser, mmrForGuild } from "./users";
 import {
   isInMatchmaking,
   addToMatchmaking,
@@ -12,29 +14,45 @@ import {
   betaWarningText,
   aboutBotText,
   enteredMatchmakingText,
+  uptimeText,
 } from "./strings";
 
 const STOP_EMOJI = Emojis.octagonal_sign;
 
+function scheduleStatusMessageUpdate(channel, messages, time = 60000) {
+  if (channel.timerId) {
+    channel.timerId();
+    channel.timerId = null;
+  }
+
+  channel.timerId = timeout(() => {
+    updateStatusMessage(channel, messages);
+  }, time);
+}
+
 async function updateStatusMessage(channel, messages) {
+  // auto-refresh every 10 minutes
+  scheduleStatusMessageUpdate(channel, messages, 1000 * 60 * 10);
   const message = messages[channel.guild.id];
   const guildId = channel.guild.id;
-  console.log("Editing status message in", guildId, channel.guild.name);
-  if (message) {
-    message.delete();
-  }
+  let messageText = "Matchmaking is online and running! ";
   const userCount = getUserCount(guildId);
   if (userCount === 0) {
-    console.log("Sending initial");
-    messages[guildId] = await channel.send(
-      `Matchmaking is online and running!\n${betaWarningText()}${instructionsText()}`
-    );
-    return;
+    messageText += `No one is playing, **be the first to queue up!**\n${uptimeText()}${betaWarningText()}${instructionsText()}`;
+  } else {
+    messageText += `Matchmaking is online and running! There are ${userCount} users playing now!\n${uptimeText()}${betaWarningText()}${instructionsText()}`;
   }
-  console.log("Sending count");
-  messages[guildId] = await channel.send(
-    `Matchmaking is online and running! There are ${userCount} users playing now!\n${betaWarningText()}${instructionsText()}`
-  );
+  messageText = messageText.trim();
+  if (message) {
+    if (channel.lastMessage.id === message.id) {
+      if (message.content !== messageText) {
+        await message.edit(messageText);
+      }
+      return;
+    }
+    message.delete();
+  }
+  messages[guildId] = await channel.send(messageText);
 }
 
 export default async function init() {
@@ -70,12 +88,32 @@ export default async function init() {
         .map((msg) => msg.delete())
     );
     await updateStatusMessage(matchmakingChannel, matchmakingStatusMessages);
+    // get away from "just now" faster?
+    scheduleStatusMessageUpdate(
+      matchmakingChannel,
+      matchmakingStatusMessages,
+      120000
+    );
   }, null);
 
   client.on("message", async (msg) => {
+    const guildId = msg.guild.id;
+    const matchmakingChannel = matchmakingChannels[guildId];
+    if (msg.channel.id !== matchmakingChannel.id) {
+      return;
+    }
+    if (msg.member.id === msg.guild.me.id) {
+      return;
+    }
+    console.log(
+      colors.yellow(msg.guild.name),
+      colors.cyan(msg.member.user.username),
+      msg.content
+    );
+
+    scheduleStatusMessageUpdate(matchmakingChannel, matchmakingStatusMessages);
+
     if (msg.content === "!matchmaking") {
-      const guildId = msg.guild.id;
-      const matchmakingChannel = matchmakingChannels[guildId];
       const state = isInMatchmaking(msg.member);
       if (state) {
         await removeFromMatchmaking(state.message.id);
@@ -92,12 +130,14 @@ export default async function init() {
       await updateStatusMessage(matchmakingChannel, matchmakingStatusMessages);
     } else if (msg.content === "!mmr") {
       const user = await updateUser(msg.member);
-      const roundedMMR = Math.round(user.mmr * 100) / 100;
+      const roundedMMR =
+        Math.round(mmrForGuild(user, msg.guild.id) * 100) / 100;
       await msg.reply(`You have ${roundedMMR} MMR.`);
       await msg.delete();
     } else if (msg.content === "!about") {
       await msg.reply(aboutBotText());
       await msg.delete();
+      await updateStatusMessage(matchmakingChannel, matchmakingStatusMessages);
     }
   });
 }
